# openapi.yaml
openapi: 3.1.0
info:
  title: Chronicler REST API
  version: 1.0.0
  description: |
    The `chronicler/v1` WordPress REST namespace: session and rule storage,
    app settings, the stateless Slack proxy, the image mirror, and character
    sheets. This is the external-client contract behind the wp-admin session
    editor and any companion app (see Gitea #112/#113).

    **Versioning policy:** changes within `chronicler/v1` are additive only —
    new routes, new optional parameters, new response fields. Anything
    breaking gets a new namespace. `info.version` tracks this document, not
    the plugin.

    **Authentication.** The core routes (slack, sessions, rules, image,
    settings, import) require a WordPress user holding the operation's
    tier: `chronicler_compose` for session/image routes and the reads the
    editor depends on, `chronicler_manage` for settings/rule/import
    writes, and `chronicler_slack_read` for the Slack proxy. The
    character-sheet routes carry their own per-route permissions, named on
    each operation: the sheet read is public, and the other three check
    WordPress capabilities on the authenticated user (`edit_post` on the
    character, `edit_others_chr_characters`, `manage_options`). External
    clients authenticate with
    [Application Passwords](https://make.wordpress.org/core/2020/11/05/application-passwords-integration-guide/)
    over HTTPS (HTTP Basic). The wp-admin bundle uses same-origin cookie +
...

...

...
servers:
...
    variables:
...
security:
  - basicAuth: []
...
paths:
  /slack/{method}:
    post:
      tags: [slack]
      operationId: slackProxy
      summary: Proxy one allowlisted Slack Web API call
      description: |
        Forwards the JSON body to the named Slack Web API method using the
        site's stored bot token and returns Slack's JSON verbatim. The
        method must be on the server's exact allowlist (conversations.list,
        conversations.history, conversations.replies, users.info,
        emoji.list, auth.test); anything else is 400. Slack 429s pass
        through with `retry_after` seconds at the top level of the body.
      parameters:
        - name: method
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z]+(\.[a-z]+)+$'
      requestBody:
        description: Arguments for the Slack method, validated server-side per method.
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Slack's JSON response, verbatim.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: The site has no Slack bot token configured (`chronicler_no_token`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Slack rate limit, passed through. Top-level `retry_after` is seconds to wait.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimited"
        "502":
          description: Slack returned an error or was unreachable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /sessions:
    get:
      tags: [sessions]
      operationId: listSessions
      summary: List sessions (light shape, no message payloads)
      description: Paginated (#164), WordPress-core style — `page`/`per_page` in, totals in the `X-WP-Total`/`X-WP-TotalPages` response headers.
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/perPage"
      responses:
        "200":
          description: One page of sessions ordered by last update, newest first.
          headers:
            X-WP-Total:
              description: Total stored sessions.
              schema:
                type: integer
            X-WP-TotalPages:
              description: Total pages at this `per_page`.
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionLight"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [sessions]
      operationId: createSession
      summary: Create a session
      description: |
        A minimal `{integration, channel, start, end}` body creates a draft;
        `rule_ids`, `editorState`, and `messages` are optional. A headless
        client can author a complete Session in one POST by including the
        transformed `messages` payload. For an open-ended session (e.g. a
        clipping bot that knows the start before the end), create with
        `end` equal to `start` and PUT the real end later.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreate"
      responses:
        "201":
          description: The stored Session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionFull"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /sessions/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [sessions]
      operationId: getSession
      summary: Fetch one session, message payload included
      responses:
        "200":
          description: The full Session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionFull"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [sessions]
      operationId: updateSession
      summary: Partially update a session
      description: Absent fields are untouched; `messages` is replaced wholesale when present.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionUpdate"
      responses:
        "200":
          description: The updated Session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionFull"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [sessions]
      operationId: deleteSession
      summary: Delete a session
      responses:
        "200":
          description: Deletion acknowledged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deleted"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /rules:
    get:
      tags: [rules]
      operationId: listRules
      summary: List rules (published only, oldest first)
      responses:
        "200":
          description: Every published rule.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [rules]
      operationId: createRule
      summary: Create a rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleCreate"
      responses:
        "201":
          description: The stored rule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /rules/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [rules]
      operationId: getRule
      summary: Fetch one rule
      responses:
        "200":
          description: The rule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [rules]
      operationId: updateRule
      summary: Partially update a rule
      description: Absent fields are untouched.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleUpdate"
      responses:
        "200":
          description: The updated rule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [rules]
      operationId: deleteRule
      summary: Delete a rule
      description: Detaches nothing — sessions keep the id in `rule_ids`; unknown ids are ignored at render time.
      responses:
        "200":
          description: Deletion acknowledged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deleted"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /image:
    get:
      tags: [media]
      operationId: image
      summary: Mirror a Slack-hosted image and redirect to the local copy
      description: |
        Mirrors the Slack image into the media library on first fetch and
        302-redirects to the local attachment URL (the default, for `<img>`
        consumers). `format=json` answers 200 with the attachment `{id,
        url}` instead — for callers that need the attachment id.
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: alt
          in: query
          schema:
            type: string
            default: ""
        - name: format
          in: query
          schema:
            type: string
            enum: [redirect, json]
            default: redirect
      responses:
        "200":
          description: The mirrored attachment (format=json only).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageAttachment"
        "302":
          description: Redirect to the local attachment URL (default).
          headers:
            Location:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          description: The download from Slack failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /settings:
    get:
      tags: [settings]
      operationId: getSettings
      summary: Fetch app settings and per-channel Session defaults
      responses:
        "200":
          description: The whole settings document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsDocument"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [settings]
      operationId: putSettings
      summary: Merge settings and channel defaults
      description: |
        Merge semantics: `settings` patches per key; `channelDefaults`
        patches per channel id — an object replaces that channel's entry,
        `null` deletes it, absent keys are kept.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsPut"
      responses:
        "200":
          description: The settings document after the merge.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsDocument"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /import:
    post:
      tags: [migration]
      operationId: import
      summary: Import the retired Node app's export payload (idempotent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportPayload"
      responses:
        "200":
          description: What the import created or updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        WordPress Application Password (HTTP Basic). Core routes require the
        operation's tier (`chronicler_compose`, `chronicler_manage`, or
        `chronicler_slack_read`, summarized above); the sheets routes name
        their own capability on each operation.
  parameters:
...
    page:
      name: page
      in: query
      description: 1-based page of results.
      schema:
        type: integer
        minimum: 1
        default: 1
...
  schemas:
...
    SlackResponse:
      type: object
      description: Slack's JSON body, verbatim.
      additionalProperties: true
...